refactor(auth): ensure correct placement of LanguageDropdown and fix lint issues
