+++ full-kit/src/components/layout/vertical-layout/vertical-layout-header.tsx
@@ -6,8 +6,8 @@ import type { DictionaryType } from "@/lib/get-dictionary"
 import type { LocaleType } from "@/types"
 
 import { SidebarTrigger } from "@/components/ui/sidebar"
+import { LanguageDropdown } from "@/components/language-dropdown"
 import { FullscreenToggle } from "@/components/layout/full-screen-toggle"
-import { LanguageDropdown } from "@/components/layout/language-dropdown"
 import { ModeDropdown } from "@/components/layout/mode-dropdown"
 import { NotificationDropdown } from "@/components/layout/notification-dropdown"
 import { UserDropdown } from "@/components/layout/user-dropdown"
